added persistence for search queries using zipcode, and set up service to check storage first for fresh results(<30 minutes old) before calling weather API

+++ Kodee_Assignment/Models/Forecast.cs
@@ -32,10 +32,10 @@ public class Weather
 
 public class Main
 {
-    public float temp { get; set; }
+    public decimal temp { get; set; }
     public int pressure { get; set; }
     public int humidity { get; set; }
-    public float temp_min { get; set; }
-    public float temp_max { get; set; }
+    public decimal temp_min { get; set; }
+    public decimal temp_max { get; set; }
 
 }
